Refactor input handling
